refactor: update comment
