+++ src/entities/workspace/stack-pipe.entity.ts
@@ -40,11 +40,6 @@ export class StackPipe extends BaseEntity {
   })
   facId: number;
 
-  @Column({
-    name: 'userid',
-  })
-  userId: string;
-
   @ManyToOne(
     () => Plant,
     plant => plant.stackPipes,
